% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{assert_valid_meta_file}
\alias{assert_valid_meta_file}
\title{Check a pData metadatad YAML file}
\usage{
assert_valid_meta_file(fn, as.list = FALSE)
}
\arguments{
\item{fn}{single character, file name}

\item{as.list}{single logical, return as list or return file name}
}
\value{
list of YAML contents or file name, controlled by `as.list`
}
\description{
Check a pData metadatad YAML file
}
