% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{custom_sample_covariates_tbl.FacileDataSet}
\alias{custom_sample_covariates_tbl.FacileDataSet}
\title{Fetches custom (user) annotations for a given user prefix}
\usage{
\method{custom_sample_covariates_tbl}{FacileDataSet}(x,
  custom_key = Sys.getenv("USER"))
}
\arguments{
\item{x}{The \code{FacileDataSet}}

\item{custom_key}{The key to use for the custom annotation}
}
\value{
covariate tbl
}
\description{
Fetches custom (user) annotations for a given user prefix
}
\seealso{
Other API: \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}
}
