% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{dataset_definitions}
\alias{dataset_definitions}
\title{Retrieves URL and description of datasets in a FacileDataSet}
\usage{
dataset_definitions(x, as.list = TRUE)
}
\arguments{
\item{as.list}{boolean, if `FALSE` (default) returns a list, otherwise
summarizes results into a tibble.}
}
\value{
meta information about the datasets in `x` as a `list` or `tibble`
}
\description{
A `FacileDataSet` can contain assay data from different "datasets" (such
as different cancer indications from the TCGA). This functions returns
description and URL information that describes these datasets in more detail,
which is specified in the FacileDataSets `meta.yaml` file.
}
