% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{default_assay.FacileDataSet}
\alias{default_assay.FacileDataSet}
\title{Retrieves the name of the default assay}
\usage{
\method{default_assay}{FacileDataSet}(x, ...)
}
\arguments{
\item{x}{A FacileDataSet}

\item{...}{dots, ignored}
}
\value{
single character, e.g. 'rnaseq'
}
\description{
Retrieves the name of the default assay
}
