% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_name_map}
\alias{feature_name_map}
\title{Returns table of names and aliases for features.}
\usage{
feature_name_map(x, feature_type)
}
\arguments{
\item{x}{\code{FacileDataSet}}

\item{feature_type}{a character vector specifying the feature type}
}
\value{
a tibble with \code{feature_id, name, type} columns, where type
  is "primary" or "alias"
}
\description{
Returns table of names and aliases for features.
}
