% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{fetch_assay_score}
\alias{fetch_assay_score}
\title{Helper function to get sample assay data from single or aggregate features}
\usage{
fetch_assay_score(x, features, samples = NULL, assay_name = NULL,
  as.matrix = FALSE, ..., subset.threshold = 700)
}
\arguments{
\item{x}{An object from a class implementing FacileInterface}

\item{features}{two-tbl of dataset and feature_id for subsetting}

\item{samples}{two-tbl of dataset and sample_id for subsetting}

\item{assay_name}{single character, e.g. rnaseq}

\item{as.matrix}{single logical, return matrix rather than long tbl?}

\item{...}{dots}

\item{subset.threshold}{single integer, take first n rows}

\item{normalized}{single logical, normalize columns?}
}
\description{
Helper function to get sample assay data from single or aggregate features
}
