% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{fetch_custom_sample_covariates}
\alias{fetch_custom_sample_covariates}
\title{Fetches and filters custom (user) annotations for a given user prefix}
\usage{
fetch_custom_sample_covariates(x, samples = NULL, covariates = NULL,
  custom_key = Sys.getenv("USER"))
}
\arguments{
\item{samples}{the facile sample descriptor}

\item{custom_key}{The key to use for the custom annotation}

\item{fds}{The \code{FacileDataSet}}
}
\value{
covariate tbl
}
\description{
Fetches and filters custom (user) annotations for a given user prefix
}
\seealso{
Other API: \code{\link{custom_sample_covariates_tbl.FacileDataSet}},
  \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{organism.FacileDataSet}}
}
\concept{API}
