% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-info.R
\name{fetch_sample_statistics}
\alias{fetch_sample_statistics}
\title{Fetch the sample statistics for sets of samples in the warehouse}
\usage{
fetch_sample_statistics(x, samples = NULL, semi = TRUE,
  assay_name = default_assay(x))
}
\arguments{
\item{x}{A \code{FacileDataSet} object}

\item{samples}{a data.frame or tbl_sqlite that has dataset and sample_id
columns}

\item{semi}{use \code{semi_join}? I've found this to be slow sometimes in
SQLite for some reason}

\item{assay_name}{parameter added to keep old API same with new "unhinged"
FacileDataSets.}
}
\value{
a tbl_df or tbl_sqlite result from the sample_stats table
}
\description{
NOTE: this function needs the axe. It has been changed to use the
assay_sample_info_tbl, but the way we handle this with the new unhinged
assay needs to change.
}
\seealso{
Other API: \code{\link{custom_sample_covariates_tbl.FacileDataSet}},
  \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_samples}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}
}
