% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{meta_file.FacileDataSet}
\alias{meta_file.FacileDataSet}
\title{Path to the meta information YAML file}
\usage{
meta_file.FacileDataSet(x)
}
\arguments{
\item{x}{A `FacileDataSet`}
}
\description{
Path to the meta information YAML file
}
\seealso{
Other FacileDataSet: \code{\link{FacileDataSet}},
  \code{\link{dbfn}}, \code{\link{hdf5fn}}
}
\concept{FacileDataSet}
