% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{meta_info.FacileDataSet}
\alias{meta_info.FacileDataSet}
\title{Retrieves the meta information for a FacileDataSet}
\usage{
\method{meta_info}{FacileDataSet}(x, fn = meta_file(x))
}
\arguments{
\item{x}{A FacileDataSet}

\item{fn}{The path to the `meta.yaml` file.}
}
\value{
The `meta.yaml` file parsed into a list-of-lists representation
}
\description{
Lots of useful information is stored in a `FacileDataSet`'s `meta.yaml` file.
This function returns all of that in a list-of-lists
}
