% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\docType{methods}
\name{organism,FacileDataSet-method}
\alias{organism,FacileDataSet-method}
\title{Get the name of the organism described by this dataset}
\usage{
\S4method{organism}{FacileDataSet}(object)
}
\arguments{
\item{object}{A FacileDataSet}
}
\value{
single character
}
\description{
Get the name of the organism described by this dataset
}
\seealso{
Other FacileInterface: \code{\link{facet_frame.FacileDataSet}}
}
\concept{FacileInterface}
