% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{organism.FacileDataSet}
\alias{organism.FacileDataSet}
\title{Retrieves the organism the data is defined over}
\usage{
organism.FacileDataSet(object)
}
\arguments{
\item{object}{A FacileDataSet}
}
\value{
`"Homo sapiens`", `"Mus musculus"`, etc.
}
\description{
A FacileDataStore is only expected to hold data for one organism.
}
\seealso{
Other API: \code{\link{custom_sample_covariates_tbl.FacileDataSet}},
  \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{samples.FacileDataSet}}
}
