% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{retrieve_samples_in_memory}
\alias{retrieve_samples_in_memory}
\title{Filter x down to specific samples}
\usage{
retrieve_samples_in_memory(criteria, cov.table = NULL)
}
\arguments{
\item{criteria}{something like a \code{tbl_sqlite} object}

\item{cov.table}{a sample descriptor}
}
\value{
filtered version of \code{x} that only has the desired samples
filter_samples <- function(x, samples=NULL) {
 join_samples(x, samples, semi=TRUE)
}
}
\description{
FacileExplorer's filter_active_samples
}
