% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{samples.FacileDataSet}
\alias{samples.FacileDataSet}
\title{#' Retrieves the sample identifiers for all samples in a FacileDataSet.
#'
#' Sample identifiers are provided as `dataset,sample_id tuples`.
#'
#' @export
#' @family API
#'
#' @param object a `FacileDataSet`
#' @return tibble with dataset and sample_id columns
#' @importFrom Biobase samples
setMethod("samples", "FacileDataSet", function(object) {
  FacileDataSet::samples.FacileDataSet(object)
})}
\usage{
\method{samples}{FacileDataSet}(x)
}
\arguments{
\item{x}{a `FacileDataSet`}
}
\value{
tibble with dataset and sample_id columns
}
\description{
#' Retrieves the sample identifiers for all samples in a FacileDataSet.
#'
#' Sample identifiers are provided as `dataset,sample_id tuples`.
#'
#' @export
#' @family API
#'
#' @param object a `FacileDataSet`
#' @return tibble with dataset and sample_id columns
#' @importFrom Biobase samples
setMethod("samples", "FacileDataSet", function(object) {
  FacileDataSet::samples.FacileDataSet(object)
})
}
\concept{FacileInterface}
