% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\docType{methods}
\name{samples,FacileDataSet-method}
\alias{samples,FacileDataSet-method}
\title{Retrieves the sample identifiers for all samples in a FacileDataSet.}
\usage{
\S4method{samples}{FacileDataSet}(object)
}
\arguments{
\item{object}{a `FacileDataSet`}
}
\value{
tibble with dataset and sample_id columns
}
\description{
Sample identifiers are provided as `dataset,sample_id tuples`.
}
\seealso{
Other API: \code{\link{custom_sample_covariates_tbl.FacileDataSet}},
  \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}
}
\concept{API}
