% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{samples.FacileDataSet}
\alias{samples.FacileDataSet}
\title{Get basic sample descriptor tibble}
\usage{
samples.FacileDataSet(object)
}
\arguments{
\item{object}{a `FacileDataSet`}
}
\value{
tibble with dataset and sample_id columns
}
\description{
Returns two-column tibble of sample_id and dataset for each sample.
}
\seealso{
Other API: \code{\link{custom_sample_covariates_tbl.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples,FacileDataSet-method}}
}
\concept{API}
